# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SanTech Data VTU Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Admin overview stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/status:
    patch:
      operationId: adminUpdateUserStatus
      tags: [admin]
      summary: Suspend or activate user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/fund:
    post:
      operationId: adminFundUser
      tags: [admin]
      summary: Manually fund user wallet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFundInput"
      responses:
        "200":
          description: Wallet funded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/transactions:
    get:
      operationId: adminGetTransactions
      tags: [admin]
      summary: List all transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /admin/data-plans:
    get:
      operationId: adminGetDataPlans
      tags: [admin]
      summary: List all data plans for pricing management
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Data plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataPlan"
    post:
      operationId: adminCreateDataPlan
      tags: [admin]
      summary: Create data plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataPlanInput"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataPlan"
  /admin/data-plans/{id}:
    patch:
      operationId: adminUpdateDataPlan
      tags: [admin]
      summary: Update data plan pricing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataPlanUpdate"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataPlan"
    delete:
      operationId: adminDeleteDataPlan
      tags: [admin]
      summary: Delete data plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [admin]
      summary: Revenue analytics by day/week/month
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: daily
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenuePoint"
  /admin/analytics/services:
    get:
      operationId: getServiceAnalytics
      tags: [admin]
      summary: Service usage analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Service analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceStat"
  /admin/notifications/broadcast:
    post:
      operationId: broadcastNotification
      tags: [admin]
      summary: Broadcast notification to all or specific users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH SCHEMAS ─────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, fullName, email, phone, role, status, emailVerified, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, admin]
        status:
          type: string
          enum: [active, suspended]
        emailVerified:
          type: boolean
        referralCode:
          type: ["string", "null"]
        referredBy:
          type: ["string", "null"]
        createdAt:
          type: string
    UserList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, suspended]

    # ─── WALLET SCHEMAS ───────────────────────────────────────────────────────
...
    AdminFundInput:
      type: object
      required: [amount, note]
      properties:
        amount:
          type: number
        note:
          type: string

    # ─── DATA SCHEMAS ─────────────────────────────────────────────────────────
    DataPlan:
      type: object
      required: [id, network, name, size, validity, price, isActive]
      properties:
        id:
          type: string
        network:
          type: string
          enum: [MTN, AIRTEL, GLO, 9MOBILE]
        name:
          type: string
        size:
          type: string
        validity:
          type: string
        price:
          type: number
        costPrice:
          type: ["number", "null"]
        isActive:
          type: boolean
    DataPlanInput:
      type: object
      required: [network, name, size, validity, price, costPrice]
      properties:
        network:
          type: string
        name:
          type: string
        size:
          type: string
        validity:
          type: string
        price:
          type: number
        costPrice:
          type: number
    DataPlanUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        costPrice:
          type: ["number", "null"]
        isActive:
          type: ["boolean", "null"]
...
    Transaction:
      type: object
      required: [id, type, status, amount, description, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [data, airtime, electricity, cable, exam, wallet_fund, wallet_transfer]
        status:
          type: string
          enum: [pending, success, failed]
        amount:
          type: number
        description:
          type: string
        reference:
          type: ["string", "null"]
        metadata:
          type: object
          additionalProperties: true
        userId:
          type: string
        createdAt:
          type: string
    TransactionList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    BroadcastInput:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
        message:
          type: string
        targetUserId:
          type: ["string", "null"]

    # ─── SUPPORT SCHEMAS ──────────────────────────────────────────────────────
...
    AdminStats:
      type: object
      required: [totalUsers, activeUsers, totalTransactions, totalRevenue, totalWalletBalance, pendingTickets, todayTransactions, todayRevenue]
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        totalTransactions:
          type: integer
        totalRevenue:
          type: number
        totalWalletBalance:
          type: number
        pendingTickets:
          type: integer
        todayTransactions:
          type: integer
        todayRevenue:
          type: number
    RevenuePoint:
      type: object
      required: [date, revenue, profit, transactions]
      properties:
        date:
          type: string
        revenue:
          type: number
        profit:
          type: number
        transactions:
          type: integer
    ServiceStat:
      type: object
      required: [service, count, revenue, percentage]
      properties:
        service:
          type: string
        count:
          type: integer
        revenue:
          type: number
        percentage:
          type: number